Added Authorization and custom requirements

+++ MotoAPI/Services/MotoService.cs
@@ -1,5 +1,9 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using MotoAPI.Authorization;
 using MotoAPI.Entitites;
 using MotoAPI.Exceptions;
 using MotoAPI.Models;
@@ -12,12 +16,27 @@ public class MotoService : IMotoService
     private readonly MotoDbContext _dbContext;
     private readonly IMapper _mapper;
     private readonly ILogger<MotoService> _logger;
+    private readonly IAuthorizationService _authorizationService;
+    private readonly IUserContextService _userContextService;
 
-    public MotoService(MotoDbContext dbContext, IMapper mapper, ILogger<MotoService> logger)
+    public MotoService(MotoDbContext dbContext, IMapper mapper, ILogger<MotoService> logger,
+        IAuthorizationService authorizationService, IUserContextService userContextService)
     {
         _dbContext = dbContext;
         _mapper = mapper;
         _logger = logger;
+        _authorizationService = authorizationService;
+        _userContextService = userContextService;
+    }
+    
+    public int Create(CreateMotoDto dto)
+    {
+        var moto = _mapper.Map<Moto>(dto);
+        moto.CreatedById = _userContextService.GetUserId;
+        _dbContext.Motos.Add(moto);
+        _dbContext.SaveChanges();
+
+        return moto.Id;
     }
 
     public void Update(int id, UpdateMotoDto dto)
@@ -29,6 +48,15 @@ public class MotoService : IMotoService
         if (moto is null)
             throw new NotFoundException("MotoShowroom not found");
 
+        var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, moto,
+                new ResourceOperationRequirement(ResourceOperation.Update))
+            .Result;
+
+        if (!authorizationResult.Succeeded)
+        {
+            throw new ForbidException();
+        }
+        
         moto.Name = dto.Name;
         moto.Description = dto.Description;
         moto.HasService = dto.HasDelivery;
@@ -36,21 +64,30 @@ public class MotoService : IMotoService
         _dbContext.SaveChanges();
     }
 
-    public void Delete(int id)
+    public void Delete(int id )
     {
         _logger.LogWarning($"MotoShowroom with id: {id} DELETE action invoked");
-        
+
         var moto = _dbContext
             .Motos
             .FirstOrDefault(m => m.Id == id);
 
-        if (moto is null) 
+        if (moto is null)
             throw new NotFoundException("MotoShowroom not found");
+
+        var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, moto,
+                new ResourceOperationRequirement(ResourceOperation.Delete))
+            .Result;
+        
+        if(!authorizationResult.Succeeded)
+        {
+            throw new ForbidException();
+        }
 
         _dbContext.Motos.Remove(moto);
         _dbContext.SaveChanges();
     }
-    
+
     public MotoDto GetById(int id)
     {
         var moto = _dbContext
@@ -63,7 +100,7 @@ public class MotoService : IMotoService
             throw new NotFoundException("MotoShowroom not found");
 
         var result = _mapper.Map<MotoDto>(moto);
-        
+
         return result;
     }
 
@@ -78,14 +115,5 @@ public class MotoService : IMotoService
         var motosDtos = _mapper.Map<List<MotoDto>>(motos);
 
         return motosDtos;
-    }
-
-    public int Create(CreateMotoDto dto)
-    {
-        var moto = _mapper.Map<Moto>(dto);
-        _dbContext.Motos.Add(moto);
-        _dbContext.SaveChanges();
-
-        return moto.Id;
     }
 }
